pyproject.toml
[tool.poetry]
name = "splortshub"
version = "0.1.0"
description = ""
authors = ["Sean Gillespie <sean@swgillespie.me>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.60.1"
uvicorn = "^0.11.8"
aiohttp = "^3.6.2"
backoff = "^1.10.0"
honeycomb-beeline = "^2.12.2"
libhoney = "^1.9.0"
asyncache = "^0.1.1"
cachetools = "^4.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.782"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
